feat Create PartPage in PartPage.js.
Created a page for each product.On the left and at the top we have Categories and Brands.Below is a block with an image of the product.Then to the right of it is the main information:Name,brand,manufacturer's number,analogues,group,article.

In the footer of the main information there is the price of the spare part, as well as a buy button, which will add the product to the cart in the future.