feat Ending developing Pages.js
The pages component is responsible for pagination, there is a main button followed by a button to return to page 1 and the next one to return to page 1 of products back. If you are already on the first page, the buttons are disabled. Then the page numbers, from 1 to 8. It is calculated based on the array of the total number of pages with products that you have filtered. And at the end, the next page button and the final page (maybe the only one), which are responsible for the opposite actions Back page.